[CLI] Move Blueprints v2 to Playground CLI package
